# docs/swagger.yaml
basePath: /api/v1/cbesuperapp/cps_action
definitions:
...
  cbe-super-app-cps-action_internal_constants_dto_productcode.ProductCodeResponse:
    properties:
      cbe_ifb_product_codes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ProductCodes'
      cbe_product_codes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ProductCodes'
      created_at:
        type: string
      id:
        type: string
      last_modified_at:
        type: string
      product_name:
        type: string
    type: object
  cbe-super-app-cps-action_internal_constants_dto_productcode.UpdateProductCodeRequest:
    properties:
      cbe_ifb_product_codes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ProductCodes'
      cbe_product_codes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ProductCodes'
      id:
        type: string
      product_name:
        type: string
    type: object
...
  cbe-super-app-cps-action_internal_constants_localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      ok:
        type: boolean
      status:
        type: integer
      timestamp:
        type: string
    type: object
...
  cbe-super-app-cps-action_internal_constants_model.ProductCode:
    properties:
      cbe_ifb_product_codes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ProductCodes'
      cbe_product_codes:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ProductCodes'
      created_at:
        type: string
      id:
        type: string
      last_modified_at:
        type: string
      service_name:
        type: string
    type: object
  cbe-super-app-cps-action_internal_constants_model.ProductCodes:
    properties:
      prd:
        type: string
      sfprd:
        type: string
      trxn:
        type: string
      vatprd:
        type: string
    type: object
...
  cbe-super-app-cps-action_internal_constants_types.PaginationMeta:
    properties:
      has_next_page:
        type: boolean
      has_prev_page:
        type: boolean
      limit:
        type: integer
      next_page:
        type: integer
      page:
        type: integer
      paging_counter:
        type: integer
      prev_page:
        type: integer
      total_docs:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handlers_rest_http_product_code.ProductCodePaginatedResponse:
    properties:
      docs:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_productcode.ProductCodeResponse'
        type: array
      meta:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.PaginationMeta'
    type: object
...
host: localhost:8080
info:
  contact:
    email: contact@eaglelionsystems.com
    name: API Support
  description: API documentation for CPS Action service
  title: CPS Action API
  version: 1.0.0
paths:
...
  /product-codes:
    get:
      description: Retrieves a paginated list of product codes. Search using service_name.Filter
        using {_id,service_name,created_at,...}
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: filter term
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers_rest_http_product_code.ProductCodePaginatedResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Fetch all product codes
      tags:
      - ProductCode
  /product-codes/{id}:
    get:
      description: Retrieves a single product code by its unique ID.
      parameters:
      - description: Product Code ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_productcode.ProductCodeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Fetch a product code by ID
      tags:
      - ProductCode
    patch:
      consumes:
      - application/json
      description: Updates an existing product code by its ID. This is a pending action
        that requires approval.
      parameters:
      - description: Product Code ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Product Code Request
        in: body
        name: productCode
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_productcode.UpdateProductCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  additionalProperties:
                    $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.ProductCode'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update a product code
      tags:
      - ProductCode
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"